requirements.txt
python-telegram-bot==20.3
aiohttp==3.9.3
python-dotenv==1.0.0
tldextract==5.1.1
regex==2023.12.25